+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add `exitCode` to return output when stream is `both`
 - Throw if `stream` is `stderr` and there is no output, you can disable this by `throwWhenEmptyStderr` option
-- THrow if `stream` is `stdout` and exit code is non-zero
+- Throw if `stream` is `stdout` and exit code is non-zero
 
 ## 1.0.5